requirements.txt
Django==4.1.2
celery==5.2.7
redis==4.3.4
flower==1.2.0
psycopg2-binary==2.9.2
djangorestframework==3.14.0
whitenoise
